Add example of using langchain to search a git repo
This commit contains an example that uses langchain to search a git
repo.

The example uses the git loader to clone a repo and then uses the
recursive character splitter to split the repo into chunks. The chunks
are then embedded using the OpenAI embeddings and stored in a Chroma
vector store. Finally, a question is asked and the vector store is
used to find the most similar documents.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>