Make children references in part1.rs
This commit makes the children field a tuble of references to Values so
that it is possible to update the gradients of the children after an
operation has been performed. Without this change the gradients of the
children would not be updated as they would be working on clones of the
values.

This required adding lifetime parameters to the Value struct which makes
it harder to understand but it is necessary to make the code work.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>