src: add "hard coded" request with result from db
This commit adds a request to the LLM when the VectorStoreTool returns a
response. The response is then used as the context for the next request
to the LLM.

This is just to provide an example and prove to myself that this was
possible. There are a number of issues with this approach I need to look
into how this is "supposed" to be implemented.
One issue is that I've not been able to find a way to specify the limit
of documents/chunks to return from the vector database. This means that
at the moment only 1 is being returned which might not provide enough
context for the LLM to answer the question.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>