src: update llm-chains-chat-demo
This commit updates the llm-chains-chat-demo to use the new gguf model
and also updates the prompts so that they now work correctly. This
commit depends on a change in llm-chain-llama where the Key Value cache
is cleared when new call to run_model is performed. If this is not done
the cache will be populated with previous results llm-chain-llama uses
the same sequence id. I'm not sure if clearning the cache is the correct
thing to do, or if we should use a different sequence id for each call
somehow.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>