src: add new batch instace for each token
This commit adds a new batch instance for each token that is decoded.
This is done mainly to avoid reusing the same batch instance which is
what the original 'simple' example in llama.cpp does.

The motivation for this is that I got feedback on the llm-chain-llama
upgrade pull request asking about this and this is a way to verify that
that is possible to do.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>