ggml: update view example using offset
This commit updates the view example to use an offset of 5. There is
something wrong with my understanding of how views work in ggml as when
I use an offset of 0 then things work as expected, that is I get a view
containing the first 5 elements of the tensor. However, when I use an
offset of 5 then I was expecting to get a view containing the last 5
elements of the tensor. This is not the case and I'm looking into why.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>