llama.cpp: add batch exploration example
This commit adds an example which is intented to be used to explore
the batch processing in llama.cpp.

The motivation for writing this is that I want to have a standalone
example that I can step through which has multiple sequences to
understand how this works better. This will probably be extended later
to understand how sbatches work too.