llama.cpp: add flake8 script
This commit adds a script to run flake8 on the llama.cpp codebase.
This is run by the llama.cpp CI pipeline but it is nice to be able to
run it locally and I'm checking this is so that I can share it between
the systems I works on.