ggml: add CUDA backend to conv2d-sched
I'm trying to reproduce and issue I'm having with a program where the
input tensor that is convoluted over seems to be modified by the conv2d
operation which is not happening in my example program here. But this
will be useful for debugging the issue and stepping though a standlone
program as the other is pretty "heavy" with regard to startup and
running.